Fail only if wkhtmltopdf exited with an error (non zero) code
Previously, we considered the conversion on the first stderr message.
This causes issues with non fatal warning, e.g. libpng warning or errors.
Now, we only treat the conversion a failure if wkhtmltopdf exits with a
non zero code.
We accumulate all the stderr messages so if wkhtmltopdf exists with an error,
we can construct a meaningful error message from them.